+++ svm-transaction/src/lib.rs
@@ -1,8 +1,6 @@
 mod instruction;
 mod message_address_table_lookup;
-mod solana_message;
-mod solana_transaction;
+mod svm_message;
+mod svm_transaction;
 
-pub use {
-    instruction::*, message_address_table_lookup::*, solana_message::*, solana_transaction::*,
-};
+pub use {instruction::*, message_address_table_lookup::*, svm_message::*, svm_transaction::*};
